Remove devicefiles properties
Device files are not a property of the radiometer object, and thus should not be enumerated as such on initialization. If the state of the system changes, i.e., a device gets added, the object property does not get updated, and thus does not reflect the state of the system anymore.
Instead, enumerate device files when necessary, i.e., inside `privateGetPortString`. This ensures that we have the most up to date state of the system when actually attempting to open a port.