Only throw warning about not using S_device and S_ambient if these fields are not empty.
If they are not empty, presumably they are not really being used, or if they are something will break that the user will be able to figure out.  The warnings were driving me nuts, given that the fields were in fact empty.

Not sure who or what is creating those empty fields in the calibration object - perhaps our OOC cal program.  In that case, we might see if we could remove them at some point altogether.